App crashes #27
The app crashes when it was trying to get the cover image by albumId, when it couldn't get any cover image it threw an exception. So, it tried to return a default cover image but it wasn't able to decode a bitmap from a resource (It only happens when use the resources ic_empty_cover or ic_app_icon unlike app_icon, so, i decided to use app_icon).